rename and move files into core
