+++ src/__tests__/analytics-pre-init.integration.test.ts
@@ -2,10 +2,10 @@ import { AnalyticsBrowser } from '..'
 import unfetch from 'unfetch'
 import { mocked } from 'ts-jest/utils'
 import { Analytics } from '../analytics'
-import { AnalyticsBuffered } from '../analytics-pre-init'
+import { AnalyticsBuffered } from '../core/buffer'
 import { Context } from '../core/context'
-import * as Factory from './test-helpers/factories'
-import { sleep } from './test-helpers/sleep'
+import * as Factory from '../test-helpers/factories'
+import { sleep } from '../test-helpers/sleep'
 import { setGlobalCDNUrl } from '../lib/parse-cdn'
 
 jest.mock('unfetch')
