+++ src/__tests__/cdn.test.ts
@@ -1,7 +1,7 @@
 import { AnalyticsBrowser } from '..'
 import { mocked } from 'ts-jest/utils'
 import unfetch from 'unfetch'
-import { createSuccess } from './test-helpers/factories'
+import { createSuccess } from '../test-helpers/factories'
 import { setGlobalCDNUrl } from '../lib/parse-cdn'
 
 jest.mock('unfetch', () => {
